feat(dashboard): update new endpoint
