+++ .changeset/icy-monkeys-deny.md
@@ -1,0 +1,5 @@
+---
+"@yukinu/dashboard": patch
+---
+
+update new endpoint
